requirements.txt
requests
pandas
openai==0.28
python-dotenv
langchain
google-cloud-aiplatform
google-generativeai
langchain-groq
streamlit
transformers
torch
